changes to recipe and add recipe page, plus image added to home, css and app changes to make variables global
